Cleanup: Remove test merge reports, update documentation, organize test files
- Remove TEST_MERGE_STATUS.md and TEST_MERGE_SUMMARY.md
- Add PROJECT_STATUS.md to .gitignore (local tracking only)
- Fix test count references (127 -> 121 tests)
- Update SECURITY_REVIEW.md and TESTING_REPORT.md
- Merge test files and remove old .extra.ts files
- Add test/helpers/ directory structure
