+++ contracts/.gitignore
@@ -27,3 +27,6 @@
 /tools/solc/
 
 deployments/localhost.json
+
+# Project status report (local tracking only, not for GitHub)
+docs/reports/PROJECT_STATUS.md
