--- contracts/test/IERC721Minimal.ts
@@ -1,5 +1,0 @@
-﻿import { expect } from "chai";
-
-describe("", function () {
-  // TODO: add tests later
-});
